Extract searchFunction used for both autocompleters
Since these were essentially the same, it made sense to re-use the
search logic amongst both.
