+++ frontend/src/app/shared/components/autocompleter/draggable-autocomplete/draggable-autocomplete.component.ts
@@ -17,6 +17,7 @@ import { merge } from 'rxjs';
 import { setBodyCursor } from 'core-app/shared/helpers/dom/set-window-cursor.helper';
 import { repositionDropdownBugfix } from 'core-app/shared/components/autocompleter/op-autocompleter/autocompleter.helper';
 import { QueryFilterResource } from 'core-app/features/hal/resources/query-filter-resource';
+import { AlternativeSearchService } from 'core-app/shared/components/work-packages/alternative-search.service';
 
 export interface DraggableOption {
   name:string;
@@ -55,8 +56,11 @@ export class DraggableAutocompleteComponent extends UntilDestroyedMixin implemen
     placeholder: this.I18n.t('js.label_add_columns'),
   };
 
-  constructor(readonly I18n:I18nService,
-    readonly dragula:DragulaService) {
+  constructor(
+    readonly I18n:I18nService,
+    readonly dragula:DragulaService,
+    readonly alternativeSearchService:AlternativeSearchService,
+  ) {
     super();
   }
 
@@ -138,36 +142,8 @@ export class DraggableAutocompleteComponent extends UntilDestroyedMixin implemen
     repositionDropdownBugfix(this.ngSelectComponent);
   }
 
-  public specialSearchStrings = {
-    done_ratio: this.I18n.t('js.work_packages.properties.done_ratio'),
-    done_ratio_alternative: this.I18n.t('js.work_packages.properties.done_ratio_alternative'),
-    work: this.I18n.t('js.work_packages.properties.work'),
-    work_alternative: this.I18n.t('js.work_packages.properties.work_alternative'),
-    remaining_work: this.I18n.t('js.work_packages.properties.remaining_work'),
-    remaining_work_alternative: this.I18n.t('js.work_packages.properties.remaining_work_alternative'),
-  };
-
   searchFunction = (term:string, currentItem:QueryFilterResource):boolean => {
-    const alternativeNames:{ [index:string]:string } = {
-      [this.specialSearchStrings.done_ratio_alternative]: this.specialSearchStrings.done_ratio,
-      [this.specialSearchStrings.work_alternative]: this.specialSearchStrings.work,
-      [this.specialSearchStrings.remaining_work_alternative]: this.specialSearchStrings.remaining_work,
-    };
-
-    const lowercaseSearchTerm = term.toLowerCase();
-    const lowercaseCurrentItemName = currentItem.name.toLowerCase();
-
-    const alternativeMatch = Object
-      .keys(alternativeNames)
-      .some((alternativeName) => {
-        return alternativeName.toLowerCase().indexOf(lowercaseSearchTerm) > -1
-          && currentItem.name === alternativeNames[alternativeName];
-      });
-
-    return (
-      lowercaseCurrentItemName.indexOf(lowercaseSearchTerm) > -1
-      || alternativeMatch
-    );
+    return this.alternativeSearchService.searchFunction(term, currentItem);
   };
 
   private updateAvailableOptions() {
